SimpleHttpExchange tests
